% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coopcredview-package.R
\docType{package}
\name{coopcredview-package}
\alias{coopcredview}
\alias{coopcredview-package}
\title{coopcredview: Visualização de Cooperativas de Crédito}
\description{
Dashboard para visualização de informações sobre as cooperativas de crédito brasileiras.
}
\author{
\strong{Maintainer}: Ricardo Theodoro \email{rtheodoro@usp.br}

}
\keyword{internal}
